datastreams: move ror transformer to common
